feat(domain): introduce domain/core.py with DomainObject and ServiceDependency; retire settings.py
- Add tiferet/domain/core.py: moves DomainObject from settings.py (no
  behavioral change) and adds the new ServiceDependency core model with
  module_path, class_name, parameters fields and get_service_type().
- Delete tiferet/domain/settings.py (retired).
- Update tiferet/domain/__init__.py: import source .settings → .core;
  add ServiceDependency to exports and __all__.
- Update all domain submodules (app, cli, di, error, feature, logging,
  request): from .settings import → from .core import.
- Rename tests/domain/test_settings.py → test_core.py; update imports
  in test_app, test_cli, test_di, test_logging; add four new
  ServiceDependency test functions.

Closes #899
