--- tests/domain/test_settings.py
@@ -1,78 +1,0 @@
-"""Tests for Tiferet Domain Settings"""
-
-# *** imports
-
-# ** infra
-import pytest
-from pydantic import Field, ValidationError
-
-# ** app
-from tiferet.domain.settings import DomainObject
-
-# *** fixtures
-
-# ** fixture: test_domain_object
-@pytest.fixture
-def test_domain_object() -> type:
-    '''
-    Fixture for a basic DomainObject subclass.
-
-    :return: The DomainObject subclass.
-    :rtype: type
-    '''
-
-    # Define a simple DomainObject subclass with one required field.
-    class TestDomainObject(DomainObject):
-        attribute: str = Field(
-            ...,
-            description='The attribute.',
-        )
-
-    # Return the class.
-    return TestDomainObject
-
-# *** tests
-
-# ** test: domain_object_construct
-def test_domain_object_construct(test_domain_object: type):
-    '''
-    Test direct construction of a DomainObject subclass.
-
-    :param test_domain_object: The DomainObject subclass to test.
-    :type test_domain_object: type
-    '''
-
-    # Construct via the standard Pydantic constructor.
-    domain_object = test_domain_object(attribute='test')
-
-    # Assert the domain object is valid.
-    assert isinstance(domain_object, test_domain_object)
-    assert domain_object.attribute == 'test'
-
-# ** test: domain_object_strict_extra_field
-def test_domain_object_strict_extra_field(test_domain_object: type):
-    '''
-    Test that DomainObject rejects unknown fields under ``extra='forbid'``.
-
-    :param test_domain_object: The DomainObject subclass to test.
-    :type test_domain_object: type
-    '''
-
-    # Constructing with an unknown field should raise.
-    with pytest.raises(ValidationError):
-        test_domain_object(attribute='test', unknown='nope')
-
-# ** test: domain_object_validate_assignment
-def test_domain_object_validate_assignment(test_domain_object: type):
-    '''
-    Test that DomainObject re-validates on attribute assignment.
-
-    :param test_domain_object: The DomainObject subclass to test.
-    :type test_domain_object: type
-    '''
-
-    # Construct and then assign an invalid type to the attribute. A list is
-    # not coercible to ``str`` even with ``coerce_numbers_to_str=True``.
-    domain_object = test_domain_object(attribute='test')
-    with pytest.raises(ValidationError):
-        domain_object.attribute = ['not', 'a', 'string']
